finalize

+++ inst/include/RcppThread/ThreadPool.hpp
@@ -185,7 +185,7 @@ ThreadPool::parallelFor(int begin, int end, F f, size_t nBatches)
         int remSize = nTasks % nBatches;
 
         for (size_t b = 0; b < nBatches; b++) {
-            size_t bSize = minSize + (remSize-- > 0);
+            int bSize = minSize + (remSize-- > 0);
             this->push([=] {
                 for (int i = begin; i < begin + bSize; ++i) {
                     f(i);
